backend/requirements.txt
fastapi
uvicorn
pydantic-ai
python-dotenv
pydantic
devtools
google-generativeai